% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_area.R
\name{density_polygons}
\alias{density_polygons}
\title{Density polygons}
\usage{
density_polygons(x, y, probs = 0.5, as_sf = FALSE, as_list = TRUE, ...)
}
\arguments{
\item{x, y}{Numeric data dimensions}

\item{probs}{Probabilities to compute density polygons for}

\item{as_sf}{Should the returned values be \link[sf:sf]{sf::sf}? Defaults to \code{FALSE}.}

\item{as_list}{Should the returned value be a list? Defaults to \code{TRUE} to
work well with tidyverse list columns}

\item{...}{Additional arguments to be passed to \code{\link[ggdensity:get_hdr]{ggdensity::get_hdr()}}}
}
\value{
A list of data frames, if \code{as_list=TRUE}, or just a data frame,
if \code{as_list=FALSE}
}
\description{
Given numeric vectors \code{x} and \code{y}, \code{density_polygons()} will return
a data frame, or list of a data frames, of the polygon defining 2d kernel
densities.
}
\details{
When using \code{density_polygons()} together with tidyverse verbs, like
\code{\link[dplyr:summarise]{dplyr::summarise()}}, \code{as_list} should be \code{TRUE}.
}
